SlidingWindow: UnexpectedResponse is no more
During tests it came to light that if the receiver is too slow in sending
responses, they may arriver after the request timed-out and retransmitted.
In this case we don’t want to fail, but simply ignore it and maybe warn the user
that the timeout might be too low.